+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 * check if the visitor IP address is valid
 * overwrite the default header parser list for the `Tracker` and allow no header usage at all
 * configure valid proxy IP address ranges for IP headers
+* added detection for Safari 15
+* updated User-Agent blacklist
+* updated referrer blacklist
+* updated referrer mapping
+* updated dependencies
 
 ## 3.8.3